requirements.txt
asttokens==2.4.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
dill==0.3.7
dnspython==2.5.0
executing==2.0.1
filelock==3.13.1
Flask==3.0.1
Flask-Cors==4.0.0
flatbuffers==23.5.26
fonttools==4.47.2
fsspec==2023.12.2
humanfriendly==10.0
idna==3.6
ipykernel==6.29.0
ipython==8.20.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
MarkupSafe==2.1.4
matplotlib==3.8.2
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.3
onnx==1.15.0
onnxruntime==1.17.0
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.0
parso==0.8.3
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.8
pure-eval==0.2.2
py-cpuinfo==9.0.0
Pygments==2.17.2
pymongo==4.6.1
pyparsing==3.1.1
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
scipy==1.12.0
seaborn==0.13.1
six==1.16.0
stack-data==0.6.3
sympy==1.12
thop==0.1.1.post2209072238
torch==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
ultralytics==8.1.4
urllib3==2.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
